pyproject.toml
[project]
name = "cnn-wave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "soundfile>=0.12",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "tqdm>=4.66",
]
